+++ src/post/post.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { PostController } from './post.controller';
+import { PostService } from './post.service';
+import { PrismaService } from '../prisma/prisma.service';
+
+@Module({
+  controllers: [PostController],
+  providers: [PostService, PrismaService]
+
+})
+export class PostModule {}
