+++ src/user/user.controller.ts
@@ -1,0 +1,25 @@
+import { 
+  Controller,
+  Get,
+  Param,
+  Post,
+  Body,
+  Put,
+  Delete,
+ } from '@nestjs/common';
+ import { UserService } from './user.service';
+ import { User as UserModel } from '@prisma/client';
+
+@Controller('user')
+export class UserController {
+  constructor(
+    private readonly userService: UserService,
+  ) {}
+
+  @Post('user')
+  async signupUser(
+    @Body() userData: { name?: string; email: string },
+  ): Promise<UserModel> {
+    return this.userService.createUser(userData);
+  }
+}
