+++ src/user/user.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { UserService } from './user.service';
+import { UserController } from './user.controller';
+import { PrismaService } from '../prisma/prisma.service';
+
+@Module({
+  controllers: [UserController,],
+  providers: [UserService, PrismaService]
+})
+export class UserModule {}
